Created authentication folder which has login.dart, forgot_password.dart, UI_utils.dart, auth_utils.dart for better readability
